backend/requirements.txt
fastapi
uvicorn
roboflow
pillow
python-dotenv
pandas
numpy
